==> 2_summary.Rmd <==
---
title: "Block-level UFP predictions - Slightly modified from the original Saha et al. 2021 Model"
author: "Magali Blanco"
date: "`r Sys.time()`"
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
    number_sections: true
    toc_float: true
    collapsed: false
    smooth_scroll: false
editor_options: 
  chunk_output_type: console
---

# Notes

UFP predictions have been updated:       

* the new predictions use recreated (similar) 1k grid UFP data to fit the model   
* predictions are for all 2010 Census blocks (like the original model) as well as for 2020 Census blocks    
* we generate predictions for all blocks (not just those with pop>0 or those within the geocovariate modeling space) since cohort members may live at highly populated locations without predictions (e.g., Manhattan)      
* to address extreme UFP predictions due to modeling outside of the covariate space, we winsorize the predictions below/above the specific quantiles
  * there are many ways to do this (e.g., could have winsorize the covariates). This is probably the most straightforward. 


```{r setup, include=FALSE}
# to run this: 
# Rscript -e 'rmarkdown::render("2_summary.Rmd", "html_document")'

knitr::opts_chunk$set(echo = F, 
                      cache=F, cache.comments = F, 
                      message = F, warning = F, 
                      tidy.opts=list(width.cutoff=60), tidy=TRUE, 
                      fig.height = 8, fig.width = 10)  

# Clear workspace of all objects and unload all extra (non-base) packages
rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
           detach, character.only=TRUE, unload=TRUE, force=TRUE))
}

pacman::p_load(tidyverse, sf,
               knitr, kableExtra,
               #ggpubr, # stat_cor() # errors out in brain
               #ggrepel, # geom_text_repel()
               tigris)  

# ggplot settings
theme_set(theme_bw())
theme_update(legend.position = "bottom")

set.seed(1)

output_path <- file.path("output", "summary", "20240716_log_truck" #"20240621_original_model"
                         )
##################################################################################################
# speed thigns up
testing_mode <- FALSE #TRUE #e.g., reduce visit designs & windows/quantile combinations

##################################################################################################
# VARIABLES
##################################################################################################
## for state outlines (doesn't impact blocks?)
yr <- 2020
project_crs <- 4269 #lat/long CRS commonly used in US 

##################################################################################################
# LOAD DATA
##################################################################################################
message("loading data")

# original model fit
lm_fit <- readRDS(file.path("original_model_fit", "lm_fit.rda"))
modeling_cov <- names(lm_fit$coefficients) %>% setdiff("(Intercept)")

## updated to have log m to truck
lm_fit_log_truck <- readRDS(file.path("original_model_fit", "lm_fit_log_truck.rda"))
modeling_cov_log_truck <- names(lm_fit_log_truck$coefficients) %>% setdiff("(Intercept)")

##################################################################################################
# model fit data
location_info <- read.csv(file.path("original_model_fit", "pnc_lur_input_1km_intra_urban.csv")) %>%
  select(data_id, city_id, site_loc, site_type, data_year)

modeling_data0 <- read.csv(file.path("original_model_fit", "input_used.csv")) %>% 
  left_join(location_info, by=c("data_id", "city_id")) %>%
  # --> New 7/16/24. how we have now modeled this
  mutate(
    # put this back on the native scale
    m_to_truck = 1/m_to_truck,
    # log transform to better specify the model. there are no 0 values
    log_m_to_truck = log(m_to_truck),
  )

modeling_data <- modeling_data0  %>%
  select(id= data_id, lat, long, #all_of(modeling_cov), 
         all_of(modeling_cov_log_truck),
         pnc=PNC) %>%
  mutate(data_type="modeling")

##################################################################################################
# quantile summary files
ufp_quantiles_2010 <- readRDS(file.path("output", "winsorize",
                                        "20240716_log_truck",
                                        "raw_quantiles_ufp_covs_2010.rds")) %>%
  mutate(year=2010)

# ufp_quantiles_2020 <- readRDS(file.path("output", "winsorize", "raw_quantiles_ufp_covs_2020.rds")) %>%
#   mutate(year=2020) 

# winsorized predictions
ufp_2010 <- readRDS(file.path("output", "predictions", "original_model", "blocks", 
                              "20240716_log_truck",
                              "modified", "ufp_covs_2010.rds")) %>%
  rename(ufp_original=ufp) %>%
  mutate(year=2010)

if(testing_mode==TRUE){
  ufp_2010 <- slice_sample(ufp_2010, n = 1e6)
  }

# ufp_2020 <- readRDS(file.path("output", "predictions", "original_model", "blocks", "modified", "ufp_covs_2020.rds"))  %>%
#   rename(ufp_original_ufp=ufp) %>%
#   mutate(year=2010)

# combine both years
ufp <- rbind(ufp_2010 #, ufp_2020
             ) %>%
  mutate(ufp_difference = ufp_winsorized-ufp_original,
         
         winsorized = ifelse(ufp_difference<0, "prediction reduced",
                                   ifelse(ufp_difference>0, "prediction increased", "not changed")
                                   )
         )
##################################################################################################
# census data
## 2013-2022 available
state_file <- file.path("output", "other", "states.rda")
if(file.exists(state_file)) {
  states <- readRDS(state_file)
  } else{
    states <- states(year = yr, cb = T #crude resolution is OK
             ) %>%
      filter(!NAME %in% c("Alaska", "Hawaii", "Puerto Rico", "Commonwealth of the Northern Mariana Islands", "Guam", "United States Virgin Islands", "District of Columbia", "American Samoa"))

    saveRDS(states, state_file)
  }

state_names <- unique(states$NAME)
if(testing_mode==TRUE){state_names <- c("New York", "Rhode Island")}

# state & county FIPS codes
data(fips_codes)

# urban areas
## include both "urbanized areas," which are densely developed areas with a population of at least 50,000, and "urban clusters," which have a population of greater than 2,500 but less than 50,000
urban_areas_file <- file.path("output", "other", "urban_areas.rda")
if(!file.exists(urban_areas_file)) {
  some_urban_areas <- urban_areas()
  saveRDS(some_urban_areas, urban_areas_file)
}

some_urban_areas <- readRDS(urban_areas_file) %>%
  filter(grepl("Urbanized Area", NAMELSAD10)) %>%
  mutate(NAME10 = gsub(", .*", "", NAME10)) %>%
  #?? some lare areas
  arrange(desc(ALAND10)) %>%
  slice(1:40)

#metro_areas <- metro_divisions()
# places_test <- places() # error
# test_statistical_area <- combined_statistical_areas()

# QC - check some predictions in Seattle
block_file <- file.path("output", "other", "king_county_blocks.rda")
if(file.exists(block_file)) {
  king_county_blocks <- readRDS(block_file)
  } else{
    king_county_blocks <- blocks(state = "WA" , county = c("King County", "Snohomish County"),year = 2010)  
    saveRDS(king_county_blocks, block_file)
  }
# test - NY
block_file <- file.path("output", "other", "new_york_blocks.rda")
if(file.exists(block_file)) {
  ny_blocks <- readRDS(block_file)
  } else{
    ny_blocks <- blocks(state = "NY", #county = c("New York"),
                        year = 2010)  
    saveRDS(ny_blocks, block_file)
  }

##################################################################################################
# COMMON VARIABLES
##################################################################################################
years <- unique(ufp$year)
all_prediction_types <- unique(ufp$winsorized) %>% setdiff(NA)

##################################################################################################
# FUNCTIONS
##################################################################################################
# function returns areas within the study area
# x.=high_ufp
# id="block_key"
# y.= filter(some_urban_areas, NAME10=="Seattle")

areas_overlap <- function(x., y., x_id) {
 
  intersects <- x. %>% 
    #st_transform(crs = utm_zone10_crs) %>%
    st_intersects(x =., y=y.) %>% 
    as.data.frame() %>%
    pull(row.id)
  
  x.[intersects,] %>% 
    select(all_of(x_id))
}


##################################################################################################

message("running analyses")

```

# covariates

covariates associated with winsorization

original model fit 
* m_to_truck has the largest weight    
  * hard to interpret since covariates are not standardized? 

```{r}
message("original model fit")
summary(lm_fit)

message("with log m_to_truck on the native scale")
summary(lm_fit_log_truck)

```


* places closer to truck routes  (m_to_truck is the inverse) are are most likely to be winsorized  

```{r}
ufp %>%
  #mutate(m_to_truck_native_scale = 1/m_to_truck) %>%
  pivot_longer(#cols = c(all_of(modeling_cov), contains("m_to_truck"))
               cols = matches(paste0(c(modeling_cov, contains("m_to_truck")), collapse = "|"))
               ) %>%
  ggplot(aes(x=value, y=ufp_difference)) +
  facet_wrap(~year+name, scales="free_x") +
  geom_smooth() +
  geom_point(aes(col=winsorized), alpha=0.3) + 
  labs(y="UFP adjustment after winsorization",
       x= "covariate value",
       col="Prediction",
       title = "predictions: covariates & winsorization"
       )

ggsave(file.path(output_path, paste0("covariates_vs_winsorization.png")), width = 16, height = 10)

# same as above just for high predictions
ufp %>%
  filter(winsorized == "prediction reduced") %>%
  #mutate(m_to_truck_native_scale = 1/m_to_truck) %>%
  pivot_longer(#cols = c(all_of(modeling_cov), contains("m_to_truck"))
                cols = matches(paste0(c(modeling_cov, contains("m_to_truck")), collapse = "|"))
    ) %>%
  ggplot(aes(x=value, y=ufp_difference)) +
  facet_wrap(~year+name, scales="free_x") +
  geom_hline(yintercept = 0, linetype=2, alpha=0.5) +
  geom_smooth(se=F) +
  geom_point(aes(col=winsorized), alpha=0.3) + 
  labs(y="UFP adjustment after winsorization",
       x= "covariate value",
       col="Prediction",
       title = "predictions reduced by covariate",
       subtitle = "predictions decreased through winsorizaiton (i.e., very high predictions)"
       )

ggsave(file.path(output_path, paste0("covariates_vs_winsorization_predictions_reduced.png")), width = 16, height = 10)

```

* **modeling data was at least 32 m from m_to_truck. prediction locations can be as close as 0 or 1 m**


```{r}
# modeling covariates & block prediction covariates
compare <- ufp %>%
  rename(id=block_key, lat=lat_block, long=long_block, pnc=ufp_original) %>%
  mutate(data_type="prediction") %>%
  select(names(modeling_data)) %>%
  rbind(modeling_data) %>%
  # # put on the native scale!
  #mutate(m_to_truck_native_scale = 1/m_to_truck) %>%
  pivot_longer(#cols = c(all_of(modeling_cov), contains("truck"))
                cols = matches(paste0(c(modeling_cov, contains("m_to_truck")), collapse = "|"))
    )

compare %>%
  group_by(name, data_type) %>%
  summarize(
    n=n(),
    min=min(value),
    median = median(value),
    mean = mean(value),
    max = max(value)
  ) %>%
  kable(caption = "distribution of modeling covariates in the modeling (observed) and prediction locations. m_to_truck is on the native scale",
        digits=2
        ) %>%
  kable_styling()

# # histogram of all covariates
# compare %>%
#   # temp for plotting 
#   mutate(value = ifelse(value==0, 1, value)) %>%
#   
#   ggplot(aes(x=value, fill=data_type)) + 
#   facet_wrap(~name, scales="free") +
#   
#   geom_density(alpha=0.4) +
#   
#   # geom_histogram(position = "dodge", ) +
#   # scale_y_log10() + 
#   
#   # geom_bar(aes(y = (..count..)/sum(..count..))) +
#   # scale_y_continuous(labels=scales::percent) +
#   
#   scale_x_log10() +
#   labs(subtitle = "x-axis is on the log10 scale",
#        y="%"
#        )
# 
# ggsave(file.path(output_path, paste0("covariate_density1.png")), width = 12, height = 10)
#  
# # histogram of m_to_truck alone
# compare %>%
#   filter(grepl("m_to_truck", name)) %>%
#   ggplot(aes(x=value, fill=data_type)) + 
#   facet_wrap(~name, scales="free") +
#   geom_histogram(position = "dodge", )  +
#   scale_x_log10() + 
#   labs(x = "value",
#        subtitle = "x-axis is on the log10 scale"
#        )

```

modeling data: PNC vs distance to truck

* things are more linear for log m_to_truck

* lu_resi has groupings by city/type of sampilng 


```{r}
modeling_data0 %>%
  select(id= data_id, city_id, site_loc, site_type, data_year, lat, long, all_of(modeling_cov), pnc=PNC) %>%  
  mutate(#m_to_truck_native_scale = 1/m_to_truck,
        m_to_truck_inverse = 1/m_to_truck,
         log_m_to_truck = log(m_to_truck),
         #m_to_truck_native_scale_log = log(m_to_truck_native_scale),
         city_id = as.factor(city_id)
         ) %>%
  pivot_longer(cols=c(all_of(modeling_cov), contains("m_to_truck"))) %>%
  
  ggplot(aes(x=value, y=pnc)) + 
  facet_wrap(~name, scales="free", switch="x") + 
  geom_point(alpha=0.3, aes(#col=city_id, #shape=site_type
                            )) +  #site_loc
  geom_smooth() + 
  geom_smooth(method="lm", se=F,col="black") +
  #stat_cor(aes(label = after_stat(r.label)))
  labs(title = "Modeling covariages vs UFP observations")

ggsave(file.path(output_path, paste0("covariates_vs_ufp_observations.png")), width = 14, height = 10)


```


ID some very high block predictions; are these on/very near roads? seattle example? 

* high predictions tend to be on major roadways

```{r}
# x_min=NA
# x_max=NA
# y_min=NA
# y_max=NA

# x_min = -122.29
# x_max = -122.2
# y_min = 47.57
# y_max = 47.61
# location="Seattle"
# predictions = c("prediction reduced")

x_min=NA
x_max=NA
y_min=NA
y_max=NA

# function maps high predictions vs block areas in a localized area
map_locations <- function(x_min=NA, x_max=NA, y_min=NA, y_max=NA, 
                          location="Seattle", block_file = king_county_blocks,
                          predictions = c("prediction reduced")
                          ) {
  
  # use entire area if not provided
  if(is.na(y_min)){y_min <- min(ufp$lat_block)}
  if(is.na(y_max)){y_max <- max(ufp$lat_block)}
  if(is.na(x_min)){x_min <- min(ufp$long_block)}
  if(is.na(x_max)){x_max <- max(ufp$long_block)}
  
  local_area <- filter(some_urban_areas, grepl(location, NAME10)) %>%
    # make into multi-point 
    summarize() %>%
    # take the outline
    st_convex_hull() 
  
  high_seattle_ufp <- ufp %>%
  filter(winsorized %in% predictions,
         # local_area can get reduced further
         (long_block >= x_min & long_block <= x_max),
         (lat_block >= y_min & lat_block <= y_max)) %>%
  st_as_sf(coords = c('long_block', 'lat_block'), crs=project_crs) %>%
  # only keep predictions in a small area
  areas_overlap(x. = .,
                y. = local_area,  
                x_id = "block_key") %>%
  left_join(ufp, by="block_key")

  # outline of the area of interest
  local_area2 <- high_seattle_ufp %>%
    # make into multi-point
    summarize() %>%
    # take the outline
    st_convex_hull()

  # blocks in the area 
  seattle_blocks <- areas_overlap(x.=block_file,
                                  #y.=local_area,
                                  y.=local_area2,
                                  x_id = "GEOID10") %>%
    left_join(st_drop_geometry(block_file))

  ggplot() + 
    geom_sf(data = seattle_blocks, fill=NA) +
    geom_sf(data = high_seattle_ufp, aes(col=ufp_original)) +
    facet_wrap(~winsorized) +
    scale_color_viridis_c(trans="log10")  +
    labs(title = "High predictions reduced through winsorization and block locations")  
  
}
##################################################################################################
# SEATTLE
############################################
# seattle - predictions reduced are mostly along/right on highways
map_locations()
ggsave(file.path(output_path, paste0("seattle_high_ufp_example1.png")), width = 8, height = 10)

# smaller areas
map_locations(x_min = -122.4, x_max = -122.2, y_min = 47.55, y_max = 47.65, 
              predictions = all_prediction_types)
ggsave(file.path(output_path, paste0("seattle_high_ufp_example2.png")), width = 10, height = 6)

## very small
map_locations(x_min = -122.29, x_max = -122.2, y_min = 47.57, y_max = 47.61, 
              predictions = all_prediction_types)
ggsave(file.path(output_path, paste0("seattle_high_ufp_example3.png")), width = 10, height = 6)

##################################################################################################
# MANHATTAN
############################################
map_locations(location = "New York", block_file = ny_blocks, 
              x_min=-74, x_max=-73.5, y_min=40.5, y_max=41.4)
ggsave(file.path(output_path, paste0("ny_high_ufp_example1.png")), width = 10, height = 6)

# closer look
map_locations(location = "New York", block_file = ny_blocks, 
              x_min=-74.00, x_max=-73.9, y_min=40.7, y_max=40.76, 
              predictions = all_prediction_types)
ggsave(file.path(output_path, paste0("ny_high_ufp_example2.png")), width = 10, height = 6)


```




# Blocks with missing predictions

these blocks have missing predictions due to missing model covariates 

```{r}
# this_year=years[1]
lapply(years, function(this_year){
  p <- ufp %>%
    filter(year==this_year) %>%
    mutate(missing_imp_a00750 = ifelse(is.na(imp_a00750), "missing imp_a00750", "UFP prediction available" )) %>%
    filter(missing_imp_a00750 == "missing imp_a00750") %>%
    st_as_sf(coords = c('long_block', 'lat_block'), crs=project_crs) %>%
    
    ggplot() + 
    geom_sf(data = states, fill=alpha(0.01)) +
    geom_sf_text(data = states, aes(label=STUSPS)) +
    geom_sf(aes(col=missing_imp_a00750), alpha=1) + 
    labs(title = "Blocks with missing covariates (i.e., predictions)",
         subtitle = this_year)
  
  ggsave(file.path(output_path, paste0("missing_predictions_", this_year, "TEST.png")), 
         plot = p, width = 16, height = 10)

  print(p)
})

```


# Quantiles used to winsorize predictions   

quantiles are based on available predictions (blocks without missing covariates) from the original UFP model predictions

We are currently winsorizing at the 1st and 99th percentiles 

```{r}
bind_rows(ufp_quantiles_2010#, 
          #ufp_quantiles_2020
          ) %>%
  filter(quantile %in% c(#0.02,0.98,
                         0.01, 0.99
                         )) %>%
  select(year, everything()) %>%
  mutate(conc = round(conc),
         year = as.character(year)) %>%
  kable(caption = "Quantiles used to winsorize year-specific predictions",
        format.args = list(big.mark=",")) %>%
  kable_styling()


```

# Distribution of UFP Predictions before & after winsorization

```{r}
ufp %>%
  select(block_key, block_year=year, contains("ufp")) %>%
  pivot_longer(cols = contains("ufp")) %>%
  mutate(block_year = as.character(block_year)) %>%
  group_by(block_year, name) %>%
  summarize(
    n = n(),
    missing = sum(is.na(value)),
    min=min(value, na.rm = T),
    Q01 = quantile(value, probs = 0.01, na.rm = T),
    Q05 = quantile(value, probs = 0.05, na.rm = T),
    Q25 = quantile(value, probs = 0.05, na.rm = T),
    Q50 = quantile(value, probs = 0.50, na.rm = T),
    Q75 = quantile(value, probs = 0.75, na.rm = T),
    Q95 = quantile(value, probs = 0.95, na.rm = T),
    Q99 = quantile(value, probs = 0.99, na.rm = T),
    max = max(value, na.rm = T)
  ) %>%
  kable(caption = "Predicted block-level PNC (pt/cm3)", 
        digits = 0, format.args = list(big.mark=",")) %>%
  kable_styling() 

```


# Comparison of original & winsorized UFP predictions 

winsorizing mostly impacts extremely high UFP predictions

```{r}
# ufp %>%
#   select(block_key, year, contains("ufp")) %>%
#   pivot_longer(cols=contains("ufp")) %>%
#   
#   ggplot(aes(x=value, fill=name)) + 
#   geom_histogram(position = "dodge") + 
#   facet_wrap(~year) +
#   labs(title = "Distribution of original and winsorized UFP predictions")

# scatterplot
ufp %>%
  ggplot(aes(x=ufp_original, y=ufp_winsorized, col=winsorized)) +
  facet_wrap(~year) +
  geom_abline(slope = 1, intercept = 0, linetype=2, alpha=0.5) +
  geom_point(alpha=0.1) + 
  labs(title = "Comparison of UFP predictions before and after winsorizing ")

```

# Prediction Maps

maps of original (pre-winsorization) UFP predictions, winsorized predictions, and prediction differences

* blocks with winsorized predictions are distributed throughout the US       
* locations where no predictions were originally generated by the original Saha model (e.g., Manhattan) are not necessarily the primary places that now have winsorized predictions. Saha et al. excluded blocks with extreme covariate values. We have winsorized the final predictions if they are extreme.          

```{r}

ufp_prediction_types <- grep("ufp", names(ufp), value = T)

# ufp_type = ufp_prediction_types[3]
lapply(ufp_prediction_types, function(ufp_type){
  
  temp <- ufp  
  
  if(ufp_type=="ufp_difference"){
    temp <- filter(ufp, winsorized != "not changed")
    }
  
  
  p <- temp %>%
    filter(year==2010) %>%
    pivot_longer(cols = contains("ufp")) %>% 
    filter(name==ufp_type) %>%  
    # 1,000s of pt/cm3
    mutate(value = value/1e3) %>%
    st_as_sf(coords = c('long_block', 'lat_block'), crs=project_crs) %>%
    
    ggplot() + 
    geom_sf(data = states, fill=alpha(0.01)) +
    # geom_sf_text(data = some_urban_areas, aes(label=NAME10), 
    #              #position = position_jitter(),
    #              #check_overlap = T, nudge_x = 1,
    #              ) +
    #geom_text_sf_repel(data = some_urban_areas, aes(label=NAME10)) +
     
    geom_sf(aes(col=value), #alpha=0.1, 
            size=0.05) + 
    facet_wrap(~name) +
    scale_color_viridis_c(trans="log10") + #scale_color_gradient2(midpoint = 30)
    labs(col="PNC (10^3 pt/cm3)")
  
  # if(ufp_type=="ufp_difference"){
  #   p <- p + scale_color_gradient2(low = "red", high = "blue")
  # }

  ggsave(file.path(output_path, paste0("predictions_", ufp_type, ".png")), 
         plot = p, width = 16, height = 10)
  
  print(p)
  
  })

```


```{r}
# # this_year=years[1]
# lapply(years, function(this_year){
#   p <- ufp %>%
#     filter(year==this_year,
#            #winsorized==TRUE
#            winsorized != "not changed"
#            ) %>%
#     st_as_sf(coords = c('long_block', 'lat_block'), crs=project_crs) %>%
#   
#     ggplot() + 
#     geom_sf(data = states, fill=alpha(0.01)) +
#     geom_sf_text(data = states, aes(label=STUSPS)) +
#     geom_sf_text(data = some_urban_areas, aes(label=NAME10), size=3, col="purple",
#                  #position = position_jitter(),
#                  check_overlap = T, #nudge_x = 1,
#                  ) +
#     geom_sf(aes(col=winsorized), alpha=0.4, size=0.2) + 
#     labs(title = "blocks with winsorized UFP predictions",
#          subtitle = this_year)
#   
#     ggsave(file.path(output_path, paste0("dichotimized_winsorized_predictions", this_year, ".png")), 
#            plot = p, width = 16, height = 10)
# 
#   print(p)
# })

```

```{r}
ufp %>%
  mutate(state_code = substr(block_key, 1,2)) %>%
  left_join(select(fips_codes, -contains("county")) %>% distinct(), by="state_code") %>% 
  group_by(year, state_name) %>%
  summarize(
    total_blocks=n(),
    proportion_winsorized = mean(winsorized != "not changed", na.rm = T)) %>%
  arrange(-proportion_winsorized) %>%
  kable(caption = "Blocks with winsorized UFP predictions", 
        digits = 2) %>%
  kable_styling() %>% 
 scroll_box(height = "500px")


```

Note: we may want to do this same winsorized location check with the actual cohort locations
